+++ CHANGELOG.md
@@ -19,6 +19,7 @@ See [STATUS.md](server/STATUS.md) to learn more about which features will remain
 - Add `POST` + `body` support for Endpoints #592
 - Refactor `Endpoint` handlers, uses a Context now #592
 - Re-build store + invite when adjusting server url #607
+- Use local atomic-server for properties and classes, improves atomic-server #604
 
 ## [v0.34.1] - 2023-02-11